is this the right way to divide up?
