+++ src/simple.rs
@@ -1,0 +1,16 @@
+
+
+named!(pub pdf_line_ending_by_macro<&[u8],&[u8]>,
+    alt!(
+        complete!(tag!(b"\r\n")) |
+        complete!(tag!(b"\r")) |
+        complete!(tag!(b"\n"))
+    )
+);
+
+
+
+#[inline]
+pub fn is_not_line_end_chars(chr: u8) -> bool {
+    (chr != b'\n' && chr != b'\r')
+}
